+++ examples/expr.rs
@@ -119,9 +119,7 @@ fn setup(mut commands: Commands, mut effects: ResMut<Assets<EffectAsset>>) {
                 gradient: size_gradient,
                 screen_space_size: false,
             })
-            .render(OrientModifier {
-                mode: OrientMode::AlongVelocity,
-            }),
+            .render(OrientModifier::new(OrientMode::AlongVelocity)),
     );
 
     commands.spawn((